refactor :polyline 구분
