+++ src/domains/Map/components/PolyLineRender.tsx
@@ -1,0 +1,45 @@
+import { Polyline } from 'react-kakao-maps-sdk';
+import { getTrafficInfo } from './getTrafficInfo';
+import type { RouteItem } from './sidebar/RoadSection';
+import type { LatLng } from '../KakaoMapContainer';
+
+function splitPathByRoad(
+  fullPath: LatLng[],
+  roadList: { distance: number; traffic_state: number }[],
+): { path: LatLng[]; traffic_state: number }[] {
+  const totalDistance = roadList.reduce((sum, r) => sum + r.distance, 0);
+  const totalPoints = fullPath.length;
+  let currentIdx = 0;
+  let accumulated = 0;
+
+  return roadList.map((r) => {
+    accumulated += r.distance;
+    const targetIdx = Math.round((accumulated / totalDistance) * totalPoints);
+    const segment = {
+      path: fullPath.slice(currentIdx, Math.max(targetIdx, currentIdx + 2)),
+      traffic_state: r.traffic_state,
+    };
+    currentIdx = targetIdx;
+    return segment;
+  });
+}
+
+export default function PolyLineRender({ route }: { route: RouteItem }) {
+  return (
+    <>
+      {splitPathByRoad(route.path, route.road).map((segment, idx) => {
+        const traffic = getTrafficInfo(segment.traffic_state);
+        return (
+          <Polyline
+            key={idx}
+            path={segment.path}
+            strokeWeight={10}
+            strokeColor={traffic.color}
+            strokeOpacity={0.8}
+            strokeStyle="solid"
+          />
+        );
+      })}
+    </>
+  );
+}
